Add files via upload
The Python file is a translation from the original fortran file. Although, the fortran file computes the response only in the frequency domain and the Python file also computes the time domain response from the frequency domain via an FFT. The results of this program are reported in the following paper:

Sánchez-Sesma, F. J. and U. Iturrarán-Viveros. Scattering and diffraction of SH waves by a finite crack: an analytical solution. Geophys. J. Int.145(3), 749-758, 2001.  

If you find the software useful, please consider to cite this paper.